requirements.txt
pandas
numpy
stable-baselines3
torch
psycopg2-binary
requests
pyyaml
matplotlib
MetaTrader5